Require login for Instagram scraping
Instagram has recently made it much more difficult to scrape for users
that are not logged in. So now logging in is required. The login data
is only kept for as long as needed to start the scrape, and deleted
immediately afterwards. Until it is, it is stored in encrypted form.