Queue chan scrape jobs automatically if configured
Fixes #92. Set `autoscrape` to `True` for a datasource in `config.py`
to automatically queue scraping jobs if these are not currently in the
queue.

Bumps version number to force updating the 8kun/8chan tables with a
'board' column (which should have been done earlier, but alas). This
is needed because these datasources are variations on the 4chan data
source that requires them.