Prevent the backend daemon from crashing
We noticed that the backend daemon would occasionally crash when it
encounters a video that has been deleted. The published date for
deleted videos returned by the bitchute api is currently 'None' which
becomes a problem for dateparse.

As a workaround to this the date will be set to something that is
obviously wrong ( 1970-01-01 00:00:00 ) before it will be parsed to fix
the basic search.

For the detailed search a video should be skipped if the returned page
contains the '404 - this page cannot be found' text.

These two changes should fix the issue mentioned in #140

Cheers