notify docker admins to consider updating their .env files
This is a proxy check since we are using the .env file copied into the Docker container. It will always be the version used to create the Docker container, but if a user already updated the .env file and for some reason has not yet used that .env file to create the 4CAT container, this message would still appear.