video_stack: move fps_mode to follow inputs, add max_video amount and enable on all video datasets, reorder overlays and comment
Since I had to change num_videos from just being dataset number of rows, I had to sort out what it ought to be and what was happening in that loop. Took a bit to wrap my head around, but hopefully its clear.

I would like to note that it actually only worked on tiktok videos precisely because the metadata file is not counted in the dataset.num_rows! The num_videos was acutally one less because it is the second to last video that applies the final overlay. We now use the actual number of videos and account for indexing from 0.