tiktokscraper do not share proxy_map
this seems to fail particularly on localhost. it seems one processor releases and then immediately reassigns localhost and so any other processors are stuck waiting. if we wanted to limit tiktok requests 4CAT wide, we could try to do that here, but we would need processors to expect to wait (instead of just failing after waiting for 600 seconds) or have some sort of in waiting queue for proxies.