4CAT-wide proxy pool (#487)
This adds a property to the worker manager, proxy_delegator, which points to an object with a number of methods for running HTTP requests concurrently and using proxies. There are some new 4CAT settings for this under the new 'Proxied HTTP requests' category.

Processors can use this feature by calling the new iterate_proxied_requests method which takes a list of URLs (or a generator or other iterable) as argument. The results are then yielded in the same order as originally requested.

The amount of allowed concurrent requests can be set on a per-proxy and a per-hostname basis. So you could configure 4CAT so that any proxy can do 50 simultaneous requests, but only up to 2 per hostname. These limits are then shared between processors. Processors all have their own request queue but proxies keep track of concurrency limits in a processor-agnostic way, i.e. two processors both iterating over requests to uva.nl will not result in twice as many concurrent requests to uva.nl (if the new method is used).

I've rewritten the 'Fetch URL Metadata' and 'Download images' processors to use this new feature - another candidate is the TikTok 'refresher' scraper.