Merge video wall and image wall (#508)
They're still two separate processors, but image wall is now basically a sub-class of the video wall.

This works because ffmpeg works with both images and videos. The neat thing is that this allows for making image walls directly from video datasets, without having to capture frames first. If you want to capture a specific frame you can still do that as an extra intermediate step though.

The image wall can now also benefit from some other new features from the video wall such as more accurate sizing and the ability to fill out the remaining space on the canvas if it's not filled by the requested amount of images.

Note that this means ffmpeg becomes a requirement to be able to use the image wall.