Notifications from notification server (#520)
Allows for remotely sourced notifications. The 'check for updates' worker already called GitHub to see if new 4CAT versions were available. It now also calls the 'phone home URL' configured in 4CAT to fetch further notifications, which can e.g. be used to inform 4CAT admins about upgrades that require extra steps (such as rebuilding Docker containers) or to send alerts about security issues.

Adds some columns to the `user_notifications_table`; a notification can now link to a page containing a longer version of its content. Remotely sourced notifications are not deleted when dismissed but just marked as 'dismissed' (and hidden from the user), to avoid a situation where the notification continuously gets added again when the notification server is queried. Once a dismissed remote notification is no longer present on the server, it is also deleted from the local database.

This can all be disabled by setting the 'phone home URL' to an empty value.